Refactor: move methods from resources into services
